+++ ams/server/src/test/resources/plugins/log_emitter.yaml
